+++ src/ValidatedNumerics.jl
@@ -37,7 +37,7 @@ export
     precedes, strictprecedes, ≺,
     entireinterval, isentire, nai, isnai, isthin, iscommon,
     widen, infimum, supremum,
-    parameters, eps, dist, 
+    parameters, eps, dist,
     pi_interval,
     midpoint_radius, interval_from_midpoint_radius,
     RoundTiesToEven, RoundTiesToAway,
@@ -45,7 +45,7 @@ export
     .., @I_str, ±
 
 export
-    displaymode
+    setdisplay
 
 export
     setindex