+++ NEWS.md
@@ -33,7 +33,7 @@ or
 
 
 ## v0.4.3
-- Fix display of intervals with different displaymode options; [#146](https://github.com/dpsanders/ValidatedNumerics.jl/pull/146)
+- Fix display of intervals with different setdisplay options; [#146](https://github.com/dpsanders/ValidatedNumerics.jl/pull/146)
 
 - Add emptyinterval(x::IntervalBox); [#145](https://github.com/dpsanders/ValidatedNumerics.jl/pull/145)
 
@@ -48,7 +48,7 @@ or
 # v0.4
 - Added decorated intervals [#112](https://github.com/dpsanders/ValidatedNumerics.jl/pull/112)
 
-- Added `displaymode` function for modifying how intervals are displayed [#115](https://github.com/dpsanders/ValidatedNumerics.jl/pull/115)
+- Added `setdisplay` function for modifying how intervals are displayed [#115](https://github.com/dpsanders/ValidatedNumerics.jl/pull/115)
 
 - Added `±` syntax for creating intervals as e.g. `1.3 ± 0.1` [#116](https://github.com/dpsanders/ValidatedNumerics.jl/pull/116)